% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\docType{data}
\name{movies}
\alias{movies}
\title{IMDB Movie Data: 1986 - 2016}
\format{A data frame with 6820 observations and 15 variables
  \describe{
   \item{budget}{Total budget (in US dollars). Budget recorded as 0 if information is missing from IMDB.com}
   \item{company}{Production company}
   \item{country}{Country of origin}
   \item{director}{Movie director}
   \item{genre}{Main genre of movie}
   \item{gross}{Total US Gross (in US dollars)}
   \item{name}{Movie title}
   \item{rating}{Movie rating (G, PG, PG-13, R, etc.)}
   \item{released}{Original release date (YYYY-MM-DD)}
   \item{runtime}{Duration of movie (in minutes)}
   \item{score}{IMDb user rating}
   \item{votes}{Number of IMDb user votes}
   \item{star}{Lead actor / actress}
   \item{writer}{Lead movie script writer}
   \item{year}{Year of original release}
}}
\source{
Data obtained from Kaggle using R \url{https://www.kaggle.com/danielgrijalvas/movies}
}
\usage{
movies
}
\description{
This data contains information on 6820 movies released in 1986 - 2016 from the Internet Movie Database (IMDb.com). The information is primarily from the "Details" and "Box Office" section on each movie's webpage. 
}
\keyword{datasets}