% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostate.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{Identifying Risk Factors of Prostate Cancer}
\format{A data frame with 67 observations and 3 variables.
  \describe{
     \item{\code{age}}{Age in years}
     \item{\code{cavol}}{Cancer volume - percentage of prostate taken up by the tumor.}
     \item{\code{psa}}{Prostate specific antigen - an antigenic enzyme released by the prostate and found in abnormally high concentrations in the blood of men with prostate cancer. Measured in nanograms per milliliter (ng/mL) of blood.}
}}
\source{
Data obtained from the "Elements of Statistical Learning" data sets webpage \url{https://web.stanford.edu/~hastie/ElemStatLearn/data.html}

Stamey, T., Kabalin, J., McNeal, J., Johnstone, I., Freiha, F., Redwine, E. and Yang, N. (1989). Prostate specific antigen in the diagnosis and treatment of adenocarcinoma of the prostate II radical prostatectomy treated patients, Journal of Urology 16: 1076–1083.
}
\usage{
prostate
}
\description{
Risk factors of prostate cancer in 67 men.
}
\keyword{datasets}
