\name{textmessages}
\alias{textmessages}
\docType{data}
\title{sd
Daily CapitalBike Share Rentals (2011-2012)
}
\description{
This data contains daily information about the rentals in the Capital Bikeshare in 2011 and 2012.
}
\usage{data("textmessages")}
\format{
  A data frame with 731 observations and 16 variables. Please see @source<https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset> for the full data dictionary.
}

\source{
Data obtained from the UCI Machine Learning Repository @source<https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset>
Original Source: @source<http://capitalbikeshare.com/system-data>
Weather Information: @source<http://www.freemeteo.com>
Holiday Schedule: @source<http://dchr.dc.gov/page/holiday-schedule>
Paper Citation: Fanaee-T, Hadi, and Gama, Joao, 'Event labeling combining ensemble detectors and background knowledge', Progress in Artificial Intelligence (2013): pp. 1-15, Springer Berlin Heidelberg
}
\keyword{sta210}
