\name{diamonds}
\alias{diamonds}
\docType{data}
\title{
Price and Characteristics of Round Cut Diamonds
}
\description{
This data set contains the price and other characteristics of about 54,000 diamonds.
\usage{data("diamonds")}
\format{
  A data frame with 53940 observations and 10 variables:
  \describe{
    \item{\code{price}}{}
    \item{\code{carat}}{}
    \item{\code{cut}}{}
    \item{\code{color}}{}
    \item{\code{clarity}}{}
    \item{\code{x}}{}
    \item{\code{y}}{}
    \item{\code{z}}{}
    \item{\code{depth}}{}
    \item{\code{table}}{}
  }
}

\source{
Data originally in the diamonds data set in the ggplot2 package.
}
\keyword{sta210}
