#' \name{textmessages}
#' \alias{textmessages}
#' \docType{data}
#' \title{ 
#' Effect of Text Messaging on Grammar Scores
#' }
#' \description{
#' This data contain the scores of 50 students on a grammar test before and after a six -month experiment. Data can be used to examine the relationship between frequent text messaging and grammar skills (as determined by a a score on a gramma test).}
#' \usage{data("textmessages")}
#' \format{
#'   A data frame with 50 observations and 3 variables.
#'   \describe{
#'    \item{\code{Group}}{Text Messagers: allowed to freely text message. \cr  Control: #' Text messaging limited during experiment}
#'    \item{\code{Baseline}}{Score on grammar test before start of experiment}
#'    \item{\code{Six_months}}{Score on grammar test at the end of the six-month experiment}
#'  }
#'}
#'
#'\source{
#'@source Data obtained from the Discovering Statistics Using R \url{https://studysites.sagepub.com/dsur/study/articles.htm}
#'\keyword{sta210}

